requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
greenlet==1.1.2
importlib-metadata==4.8.1
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.26
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
zipp==3.6.0
